+++ solc/CommandLineInterface.cpp
@@ -607,7 +607,7 @@ Allowed options)",
 		(
 			g_strEVMVersion.c_str(),
 			po::value<string>()->value_name("version"),
-			"Select desired EVM version. Either homestead, tangerineWhistle, spuriousDragon, byzantium (default) or constantinople."
+			"Select desired EVM version. Either homestead, tangerineWhistle, spuriousDragon, byzantium (default), constantinople or petersburg."
 		)
 		(g_argOptimize.c_str(), "Enable bytecode optimizer.")
 		(